requirements.txt
streamlit
easyocr
groq
psycopg2-binary
numpy
scikit-learn
pdfplumber
pillow
python-dotenv